[PROJECT-WAR][1.0.0-SNAPSHOT][류강하] [BT] SP-80 #comment 1. 이메일 삭제 시 파라미터 값 검증 추가  2. 동일한 타이틀을 가진 노드 조회 메서드 추가 [RV] +review SR @spear79

+++ trunk/development-standard/standard-projects/project-war/src/main/java/standard/mvc/component/business/community/newsletter/controller/NewsletterController.java
@@ -48,10 +48,10 @@ import egovframework.com.ext.jstree.springiBatis.core.vo.ComprehensiveTree;
  *  -------        ------------   -----------------------
  *  2014. 12. 09.  류강하                 최초 생성
  *  2014. 12. 13.  전경훈                 Email 검증부분 추가 및 jstree 필수 파라미터 설정
- *  2014. 12. 15.  류강하                 getChildNode 추가
- *  2014. 12. 19.  류강하                 사용자 및 관리자 화면을 동시 처리하게끔 addNode 보완, removeNode 추가
- *  2014. 12. 20.  류강하                 alterNode, alterNodeType 추가
- *  2014. 12. 28.  류강하                 searchNode, moveNode 추가
+ *  2014. 12. 15.  류강하                 getChildNode() 추가
+ *  2014. 12. 19.  류강하                 사용자 및 관리자 화면을 동시 처리하게끔 addNode() 보완, removeNode() 추가
+ *  2014. 12. 20.  류강하                 alterNode(), alterNodeType() 추가
+ *  2014. 12. 28.  류강하                 searchNode(), moveNode() 추가
  *  2015. 01. 03   류강하                 사용자와 관리자를 위한 로직을 분리하기 위해 관리자 관련 메서드들을 NewsletterAdminController로 옮김.
  * 
  *  Copyright (C) 2014 by 313 DeveloperGroup  All right reserved.
@@ -111,8 +111,23 @@ public class NewsletterController extends GenericAbstractController {
      */
     @ResponseBody
     @RequestMapping("/removeEmail.do")
-    public int removeNode(@ModelAttribute NewsletterComprehensiveTree newsletterComprehensiveTree) 
+    public int removeNode(@ModelAttribute NewsletterComprehensiveTree newsletterComprehensiveTree
+                                        , HttpServletRequest request) 
             throws Exception {
+        
+        String c_id = request.getParameter("c_id");
+        
+        if (c_id == null) {
+            throw new RuntimeException("removeNode()'s essential parameter not set.");
+        }
+        
+        if (!c_id.matches("\\d+")) {
+            throw new RuntimeException("removeNode()'s c_id parameter value is not valid.");
+        }
+        
+        if (Integer.parseInt(c_id) < 4) {
+            throw new RuntimeException("removeNode()'s c_id parameter value is predefined value.");
+        }
         
         return newsletterService.removeNode(newsletterComprehensiveTree);
     }
